Add role config in twenty-sdk
