+++ packages/twenty-sdk/src/application/role-config.ts
@@ -1,0 +1,11 @@
+import type { Role } from 'twenty-shared/application';
+
+export type RoleConfig = Omit<
+  Role,
+  | 'canAccessAllTools'
+  | 'canUpdateAllSettings'
+  | 'canBeAssignedToAgents'
+  | 'canBeAssignedToUsers'
+  | 'canBeAssignedToApiKeys'
+  | 'canBeAssignedToApplications'
+>;
